+++ aiobotocore/credentials.py
@@ -1,11 +1,8 @@
 import asyncio
-import datetime
-import json
 import logging
 import os
 import subprocess
 from copy import deepcopy
-from hashlib import sha1
 
 import botocore.compat
 from botocore import UNSIGNED
@@ -27,6 +24,7 @@ from botocore.credentials import (
     CredentialResolver,
     CredentialRetrievalError,
     Credentials,
+    DeferredRefreshableCredentials,
     EnvProvider,
     InstanceMetadataProvider,
     InvalidConfigError,
@@ -39,6 +37,7 @@ from botocore.credentials import (
     RefreshableCredentials,
     RefreshWithMFAUnsupportedError,
     SharedCredentialProvider,
+    SSOCredentialFetcher,
     SSOProvider,
     SSOTokenLoader,
     UnauthorizedSSOTokenError,
@@ -50,7 +49,6 @@ from botocore.credentials import (
     parse,
     resolve_imds_endpoint_mode,
 )
-from dateutil.tz import tzutc
 
 from aiobotocore._helpers import resolve_awaitable
 from aiobotocore.config import AioConfig
@@ -364,7 +362,9 @@ class AioRefreshableCredentials(RefreshableCredentials):
         return self._frozen_credentials
 
 
-class AioDeferredRefreshableCredentials(AioRefreshableCredentials):
+class AioDeferredRefreshableCredentials(
+    DeferredRefreshableCredentials, AioRefreshableCredentials
+):
     def __init__(self, refresh_using, method, time_fetcher=_local_now):
         self._refresh_using = refresh_using
         self._access_key = None
@@ -375,11 +375,6 @@ class AioDeferredRefreshableCredentials(AioRefreshableCredentials):
         self._refresh_lock = asyncio.Lock()
         self.method = method
         self._frozen_credentials = None
-
-    def refresh_needed(self, refresh_in=None):
-        if self._frozen_credentials is None:
-            return True
-        return super().refresh_needed(refresh_in)
 
 
 class AioCachedCredentialFetcher(CachedCredentialFetcher):
@@ -747,11 +742,14 @@ class AioAssumeRoleProvider(AssumeRoleProvider):
     async def _resolve_credentials_from_profile(self, profile_name):
         profiles = self._loaded_config.get('profiles', {})
         profile = profiles[profile_name]
-
         if (
             self._has_static_credentials(profile)
             and not self._profile_provider_builder
         ):
+            # This is only here for backwards compatibility. If this provider
+            # isn't given a profile provider builder we still want to be able
+            # handle the basic static credential case as we would before the
+            # provile provider builder parameter was added.
             return self._resolve_static_credentials_from_profile(profile)
         elif self._has_static_credentials(
             profile
@@ -770,7 +768,6 @@ class AioAssumeRoleProvider(AssumeRoleProvider):
                     error_msg=error_message % profile_name,
                 )
             return credentials
-
         return await self._load_creds_via_assume_role(profile_name)
 
     def _resolve_static_credentials_from_profile(self, profile):
@@ -971,52 +968,9 @@ class AioCredentialResolver(CredentialResolver):
         return None
 
 
-class AioSSOCredentialFetcher(AioCachedCredentialFetcher):
-    _UTC_DATE_FORMAT = '%Y-%m-%dT%H:%M:%SZ'
-
-    def __init__(
-        self,
-        start_url,
-        sso_region,
-        role_name,
-        account_id,
-        client_creator,
-        token_loader=None,
-        cache=None,
-        expiry_window_seconds=None,
-        token_provider=None,
-        sso_session_name=None,
-    ):
-        self._client_creator = client_creator
-        self._sso_region = sso_region
-        self._role_name = role_name
-        self._account_id = account_id
-        self._start_url = start_url
-        self._token_loader = token_loader
-        self._token_provider = token_provider
-        self._sso_session_name = sso_session_name
-        super().__init__(cache, expiry_window_seconds)
-
-    def _create_cache_key(self):
-        args = {
-            'roleName': self._role_name,
-            'accountId': self._account_id,
-        }
-        if self._sso_session_name:
-            args['sessionName'] = self._sso_session_name
-        else:
-            args['startUrl'] = self._start_url
-
-        args = json.dumps(args, sort_keys=True, separators=(',', ':'))
-        argument_hash = sha1(args.encode('utf-8')).hexdigest()
-        return self._make_file_safe(argument_hash)
-
-    def _parse_timestamp(self, timestamp_ms):
-        # fromtimestamp expects seconds so: milliseconds / 1000 = seconds
-        timestamp_seconds = timestamp_ms / 1000.0
-        timestamp = datetime.datetime.fromtimestamp(timestamp_seconds, tzutc())
-        return timestamp.strftime(self._UTC_DATE_FORMAT)
-
+class AioSSOCredentialFetcher(
+    SSOCredentialFetcher, AioCachedCredentialFetcher
+):
     async def _get_credentials(self):
         """Get credentials by calling SSO get role credentials."""
         config = Config(